MyRoutine/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.4
django-multiselectfield==0.1.12
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
